자유게시판 댓글기능

+++ src/main/java/com/culture/repository/BoardFreeReplyRepository.java
@@ -4,6 +4,6 @@ import com.culture.entity.boardFree.BoardFreeReply;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.querydsl.QuerydslPredicateExecutor;
 
-public interface BoardFreeReplyRepository extends JpaRepository<BoardFreeReply, Integer>,
+public interface BoardFreeReplyRepository extends JpaRepository<BoardFreeReply, Long>,
         QuerydslPredicateExecutor<BoardFreeReply>, BoardFreeReplyRepositoryCustom {
 }
